+++ lib/src/core/lifecycle/live_data_scope.dart
@@ -3,7 +3,7 @@ import 'live_data.dart';
 mixin LiveDataScope {
   final List<LiveData> _liveDataList = [];
 
-  void addLiveData(LiveData liveData) {
+  void join(LiveData liveData) {
     if (!_liveDataList.contains(liveData)) {
       _liveDataList.add(liveData);
     }
@@ -11,9 +11,8 @@ mixin LiveDataScope {
 
   /// 创建 ViewModelState
   /// notify 创建时是否更新状态
-  LiveData<M> useState<M>(M value, {bool notify = false}) {
-    var liveData = LiveData.useState(value, notify: notify);
-    addLiveData(liveData);
+  LiveData<T> useState<T>(T value, [bool notify = false]) {
+    var liveData = LiveData.useState(value, this, notify);
     return liveData;
   }